[🚚Rename]: 회원가입 페이지 체크박스 파일 폴더 이동 1. 파일 위치 바꾼거 경로 이동. 2. 경로 수정
